Misc. code changes
Changes based on enabling -Weverything (all warnings)

• Removed few unused macros
• Replaced some argument types with “id” because of type mismatch
(table view delegates)
• Modified TXHumanReadableTimeInterval() to use NSInvocation instead of
objc_msgSend because the return value is an NSInteger
• Fixed some numbers being cast as floats when they are supposed to be
doubles
• Fixed IRCChannelConfig returning self.secretValue (keychain value)
when its supposed to return _secretKey (temporary value in memory)
• Fixed [IRCUser presentAwayMessageFor301] performing incorrect
comparison

+++ Classes/Dialogs/TDCServerHighlightListSheet.m
@@ -116,7 +116,7 @@
 
 		CGFloat calculatedTextHeight = [[entryItem renderedMessage] pixelHeightInWidth:(NSWidth(textFieldFrame) - (_renderedMessageTextFieldLeftRightPadding * 2))];
 
-		CGFloat finalRowHeight = (ceilf(calculatedTextHeight / [aTableView rowHeight]) * [aTableView rowHeight]);
+		CGFloat finalRowHeight = (ceil(calculatedTextHeight / [aTableView rowHeight]) * [aTableView rowHeight]);
 
 		[entryItem setRowHeight:finalRowHeight];
